requirements.txt
select
socket
sys
pycrypto
ast
base64
_thread
